transfer quarantine

+++ src/main/java/org/karnak/backend/dicom/ForwardUtil.java
@@ -32,6 +32,7 @@ import org.dcm4che3.net.DataWriterAdapter;
 import org.dcm4che3.net.InputStreamDataWriter;
 import org.dcm4che3.net.PDVInputStream;
 import org.dcm4che3.net.Status;
+import org.karnak.backend.model.Quarantine;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.weasis.core.util.FileUtil;
@@ -42,6 +43,7 @@ import org.weasis.dicom.param.AttributeEditorContext;
 import org.weasis.dicom.param.AttributeEditorContext.Abort;
 import org.weasis.dicom.param.ConnectOptions;
 import org.weasis.dicom.param.DicomNode;
+import org.weasis.dicom.param.DicomProgress;
 import org.weasis.dicom.util.ServiceUtil;
 import org.weasis.dicom.util.ServiceUtil.ProgressStatus;
 import org.weasis.dicom.util.StoreFromStreamSCU;
@@ -275,6 +277,7 @@ public class ForwardUtil {
         }
 
         if (context.getAbort() == Abort.FILE_EXCEPTION) {
+          sourceNode.setQuarantine(new Quarantine(copy, p));
           if (p.getData() instanceof PDVInputStream) {
             ((PDVInputStream) p.getData()).skipAll();
           }
@@ -299,7 +302,7 @@ public class ForwardUtil {
       progressNotify(
           destination, p.getIuid(), p.getCuid(), true, streamSCU.getNumberOfSuboperations());
       if (e.getAbort() == Abort.CONNECTION_EXCEPTION) {
-        transferQuarantine(sourceNode, copy, p);
+        sourceNode.setQuarantine(new Quarantine(copy, p));
         throw e;
       }
     } catch (Exception e) {
@@ -308,7 +311,7 @@ public class ForwardUtil {
       }
       progressNotify(
           destination, p.getIuid(), p.getCuid(), true, streamSCU.getNumberOfSuboperations());
-      transferQuarantine(sourceNode, copy, p);
+      sourceNode.setQuarantine(new Quarantine(copy, p));
       LOGGER.error(ERROR_WHEN_FORWARDING, e);
     } finally {
       streamSCU.triggerCloseExecutor();
@@ -398,7 +401,7 @@ public class ForwardUtil {
       progressNotify(
           destination, p.getIuid(), p.getCuid(), true, streamSCU.getNumberOfSuboperations());
       if (e.getAbort() == Abort.CONNECTION_EXCEPTION) {
-        transferQuarantine(fwdNode, copy, p);
+        fwdNode.setQuarantine(new Quarantine(copy, p));
         throw e;
       }
     } catch (Exception e) {
@@ -407,7 +410,7 @@ public class ForwardUtil {
       }
       progressNotify(
           destination, p.getIuid(), p.getCuid(), true, streamSCU.getNumberOfSuboperations());
-      transferQuarantine(fwdNode, copy, p);
+      fwdNode.setQuarantine(new Quarantine(copy, p));
       LOGGER.error(ERROR_WHEN_FORWARDING, e);
     } finally {
       streamSCU.triggerCloseExecutor();
@@ -448,13 +451,13 @@ public class ForwardUtil {
           if (p.getData() instanceof PDVInputStream) {
             ((PDVInputStream) p.getData()).skipAll();
           }
-          transferQuarantine(fwdNode, copy, p);
+          fwdNode.setQuarantine(new Quarantine(copy, p));
           throw new AbortException(context.getAbort(), context.getAbortMessage());
         } else if (context.getAbort() == Abort.CONNECTION_EXCEPTION) {
           if (p.getAs() != null) {
             p.getAs().abort();
           }
-          transferQuarantine(fwdNode, copy, p);
+          fwdNode.setQuarantine(new Quarantine(copy, p));
           throw new AbortException(
               context.getAbort(), "STOW-RS abort: " + context.getAbortMessage());
         }
@@ -471,12 +474,12 @@ public class ForwardUtil {
     } catch (AbortException e) {
       progressNotify(destination, p.getIuid(), p.getCuid(), true, 0);
       if (e.getAbort() == Abort.CONNECTION_EXCEPTION) {
-        transferQuarantine(fwdNode, copy, p);
+        fwdNode.setQuarantine(new Quarantine(copy, p));
         throw e;
       }
     } catch (Exception e) {
       progressNotify(destination, p.getIuid(), p.getCuid(), true, 0);
-      transferQuarantine(fwdNode, copy, p);
+      fwdNode.setQuarantine(new Quarantine(copy, p));
       LOGGER.error(ERROR_WHEN_FORWARDING, e);
     } finally {
       files = cleanOrGetBulkDataFiles(in, copy == null);
@@ -499,10 +502,10 @@ public class ForwardUtil {
         editors.forEach(e -> e.apply(attributes, context));
 
         if (context.getAbort() == Abort.FILE_EXCEPTION) {
-          transferQuarantine(fwdNode, copy, p);
+          fwdNode.setQuarantine(new Quarantine(copy, p));
           throw new AbortException(context.getAbort(), context.getAbortMessage());
         } else if (context.getAbort() == Abort.CONNECTION_EXCEPTION) {
-          transferQuarantine(fwdNode, copy, p);
+          fwdNode.setQuarantine(new Quarantine(copy, p));
           throw new AbortException(
               context.getAbort(), "DICOM associtation abort. " + context.getAbortMessage());
         }
@@ -518,12 +521,12 @@ public class ForwardUtil {
       }
     } catch (HttpException httpException) {
       progressNotify(destination, p.getIuid(), p.getCuid(), true, 0);
-      transferQuarantine(fwdNode, copy, p);
+      fwdNode.setQuarantine(new Quarantine(copy, p));
       throw new AbortException(Abort.FILE_EXCEPTION, "DICOMWeb forward", httpException);
     } catch (AbortException e) {
       progressNotify(destination, p.getIuid(), p.getCuid(), true, 0);
       if (e.getAbort() == Abort.CONNECTION_EXCEPTION) {
-        transferQuarantine(fwdNode, copy, p);
+        fwdNode.setQuarantine(new Quarantine(copy, p));
         throw e;
       }
     } catch (Exception e) {
@@ -556,8 +559,8 @@ public class ForwardUtil {
     return UID.ImplicitVRLittleEndian;
   }
 
-  public static void transferQuarantine(ForwardDicomNode fwdNode, Attributes quarantineData, Params p) {
-    try{
+  public static void transferQuarantine(ForwardDicomNode fwdNode, Params p) {
+    try {
       final DicomNode quarantineNode = new DicomNode("NODE2", "localhost", 4445);
       AdvancedParams params = new AdvancedParams();
       ConnectOptions connectOptions = new ConnectOptions();
@@ -568,17 +571,29 @@ public class ForwardUtil {
       connectOptions.setMaxOpsPerformed(50);
       params.setConnectOptions(connectOptions);
 
+      DicomProgress progress = new DicomProgress();
+
+      List<AttributeEditor> editors = new ArrayList<>();
       DicomForwardDestination quarantineDestination =
-          new DicomForwardDestination(null, params, fwdNode, quarantineNode, false, null, null);
+          new DicomForwardDestination(null, params, fwdNode, quarantineNode, true, progress, editors);
 
-      List<File> files = transfer(fwdNode, quarantineDestination, quarantineData, p);
+      Quarantine quarantine = fwdNode.getQuarantine();
+      if (quarantine != null) {
+        Attributes quarantineAttributes = quarantine.getQuanrantineAttribute();
+        Params p2 = quarantine.getP();
+        prepareTransfer(quarantineDestination, p2);
+        List<File> files = transfer(fwdNode, quarantineDestination, quarantineAttributes, p2);
 
-      if (files != null) {
-        // Force to clean if tmp bulk files
-        for (File file : files) {
-          FileUtil.delete(file);
+        if (files != null) {
+          // Force to clean if tmp bulk files
+          for (File file : files) {
+            FileUtil.delete(file);
+          }
         }
+      } else {
+        storeOneDestination(fwdNode, quarantineDestination, p);
       }
+
     } catch (Exception e) {
       LOGGER.error("Error when forwarding in quarantine", e);
     }
